This is a major commit with the code for many new features including automatic DELZN insertion, support for the digital backends and much more.  See the manual notes for more details.  I hope to do at least a beta release soon, maybe even a full release.
git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@322 d0039845-cc32-0410-9ae4-e027caa9de09

+++ src/Sched/fsvlba.f
@@ -11,7 +11,7 @@ C
 C     Common for communication with VLBASU.
 C
       INTEGER  MBBCFREQ
-      REAL     LBBCFREQ(MCHAN)
+      DOUBLE PRECISION  LBBCFREQ(MCHAN)
       COMMON  / CVLBA /  LBBCFREQ, MBBCFREQ
       SAVE    /CVLBA/
 C ----------------------------------------------------------------------
